// apps/backends/postgres/migrations/snapshots/54e458ace6b116150d56f24344d03318dfc6746ae5ea0541d72db660eeadfa59/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly AddressSchema: {
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly userId: CodecTypes['pg/uuid@1']['output'];
      readonly fullName: CodecTypes['pg/text@1']['output'];
      readonly address: CodecTypes['pg/text@1']['output'];
      readonly state: CodecTypes['pg/text@1']['output'];
      readonly city: CodecTypes['pg/text@1']['output'];
      readonly country: CodecTypes['pg/text@1']['output'];
      readonly postalCode: CodecTypes['pg/text@1']['output'];
      readonly isDefault: CodecTypes['pg/bool@1']['output'];
    };
    readonly Brand: {
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
    };
    readonly Category: {
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
    };
    readonly Session: {
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly userId: CodecTypes['pg/uuid@1']['output'];
      readonly userAgent: CodecTypes['pg/text@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly expiresAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly refreshJti: CodecTypes['pg/text@1']['output'];
      readonly prevRefreshJti: CodecTypes['pg/text@1']['output'] | null;
      readonly rotatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'] | null;
    };
    readonly SubCategory: {
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly categoryId: CodecTypes['pg/uuid@1']['output'];
    };
    readonly User: {
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly passwordHash: CodecTypes['pg/text@1']['output'] | null;
      readonly name: CodecTypes['pg/text@1']['output'] | null;
      readonly googleId: CodecTypes['pg/text@1']['output'] | null;
      readonly authProvider: 'local' | 'google';
      readonly avatar: CodecTypes['pg/text@1']['output'] | null;
      readonly role: 'user' | 'admin';
      readonly points: CodecTypes['pg/int4@1']['output'];
      readonly verified: CodecTypes['pg/bool@1']['output'];
    };
    readonly VerificationLink: {
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly type: CodecTypes['pg/text@1']['output'];
      readonly token: CodecTypes['pg/text@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly expiresAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly AddressSchema: {
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly userId: CodecTypes['pg/uuid@1']['input'];
      readonly fullName: CodecTypes['pg/text@1']['input'];
      readonly address: CodecTypes['pg/text@1']['input'];
      readonly state: CodecTypes['pg/text@1']['input'];
      readonly city: CodecTypes['pg/text@1']['input'];
      readonly country: CodecTypes['pg/text@1']['input'];
      readonly postalCode: CodecTypes['pg/text@1']['input'];
      readonly isDefault: CodecTypes['pg/bool@1']['input'];
    };
    readonly Brand: {
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
    };
    readonly Category: {
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
    };
    readonly Session: {
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly userId: CodecTypes['pg/uuid@1']['input'];
      readonly userAgent: CodecTypes['pg/text@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly expiresAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly refreshJti: CodecTypes['pg/text@1']['input'];
      readonly prevRefreshJti: CodecTypes['pg/text@1']['input'] | null;
      readonly rotatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'] | null;
    };
    readonly SubCategory: {
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly categoryId: CodecTypes['pg/uuid@1']['input'];
    };
    readonly User: {
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly passwordHash: CodecTypes['pg/text@1']['input'] | null;
      readonly name: CodecTypes['pg/text@1']['input'] | null;
      readonly googleId: CodecTypes['pg/text@1']['input'] | null;
      readonly authProvider: 'local' | 'google';
      readonly avatar: CodecTypes['pg/text@1']['input'] | null;
      readonly role: 'user' | 'admin';
      readonly points: CodecTypes['pg/int4@1']['input'];
      readonly verified: CodecTypes['pg/bool@1']['input'];
    };
    readonly VerificationLink: {
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly type: CodecTypes['pg/text@1']['input'];
      readonly token: CodecTypes['pg/text@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly expiresAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly addressSchema: {
      readonly address: CodecTypes['pg/text@1']['output'];
      readonly city: CodecTypes['pg/text@1']['output'];
      readonly country: CodecTypes['pg/text@1']['output'];
      readonly fullName: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly isDefault: CodecTypes['pg/bool@1']['output'];
      readonly postalCode: CodecTypes['pg/text@1']['output'];
      readonly state: CodecTypes['pg/text@1']['output'];
      readonly userId: CodecTypes['pg/uuid@1']['output'];
    };
    readonly brand: {
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
    };
    readonly category: {
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
    };
    readonly session: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly expiresAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly prevRefreshJti: CodecTypes['pg/text@1']['output'] | null;
      readonly refreshJti: CodecTypes['pg/text@1']['output'];
      readonly rotatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'] | null;
      readonly userAgent: CodecTypes['pg/text@1']['output'];
      readonly userId: CodecTypes['pg/uuid@1']['output'];
    };
    readonly subCategory: {
      readonly categoryId: CodecTypes['pg/uuid@1']['output'];
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
    };
    readonly user: {
      readonly authProvider: 'local' | 'google';
      readonly avatar: CodecTypes['pg/text@1']['output'] | null;
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly googleId: CodecTypes['pg/text@1']['output'] | null;
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'] | null;
      readonly passwordHash: CodecTypes['pg/text@1']['output'] | null;
      readonly points: CodecTypes['pg/int4@1']['output'];
      readonly role: 'user' | 'admin';
      readonly verified: CodecTypes['pg/bool@1']['output'];
    };
    readonly verificationLink: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly expiresAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly token: CodecTypes['pg/text@1']['output'];
      readonly type: CodecTypes['pg/text@1']['output'];
    };
  };
};
export type StorageColumnInputTypes = {
  readonly public: {
    readonly addressSchema: {
      readonly address: CodecTypes['pg/text@1']['input'];
      readonly city: CodecTypes['pg/text@1']['input'];
      readonly country: CodecTypes['pg/text@1']['input'];
      readonly fullName: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly isDefault: CodecTypes['pg/bool@1']['input'];
      readonly postalCode: CodecTypes['pg/text@1']['input'];
      readonly state: CodecTypes['pg/text@1']['input'];
      readonly userId: CodecTypes['pg/uuid@1']['input'];
    };
    readonly brand: {
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
    };
    readonly category: {
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
    };
    readonly session: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly expiresAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly prevRefreshJti: CodecTypes['pg/text@1']['input'] | null;
      readonly refreshJti: CodecTypes['pg/text@1']['input'];
      readonly rotatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'] | null;
      readonly userAgent: CodecTypes['pg/text@1']['input'];
      readonly userId: CodecTypes['pg/uuid@1']['input'];
    };
    readonly subCategory: {
      readonly categoryId: CodecTypes['pg/uuid@1']['input'];
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
    };
    readonly user: {
      readonly authProvider: 'local' | 'google';
      readonly avatar: CodecTypes['pg/text@1']['input'] | null;
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly googleId: CodecTypes['pg/text@1']['input'] | null;
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'] | null;
      readonly passwordHash: CodecTypes['pg/text@1']['input'] | null;
      readonly points: CodecTypes['pg/int4@1']['input'];
      readonly role: 'user' | 'admin';
      readonly verified: CodecTypes['pg/bool@1']['input'];
    };
    readonly verificationLink: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly expiresAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly token: CodecTypes['pg/text@1']['input'];
      readonly type: CodecTypes['pg/text@1']['input'];
    };
  };
};
// ...
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;